pyproject.toml
[project]
name = "full-duplex-voice-ai"
version = "0.1.0"
description = "Real-time low-latency full-duplex audio streaming server"
requires-python = ">=3.11"
dependencies = [
    "aiortc>=1.14.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "numpy>=2.0.0",
    "soxr>=0.5.0",
    "faster-whisper>=1.0.0",
    "onnxruntime>=1.17.0",
    "nvidia-cublas-cu12>=12.0",
    "nvidia-cudnn-cu12>=9.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
